damn

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/Calculate/BinarySearch.java
@@ -8,7 +8,7 @@ public class BinarySearch {
     public BinarySearch(double[] data){
         this.D = data;
     }
-    public double find(int left ,int right, double goal) {
+    public int find(int left ,int right, double goal) {
         int mid = (int) Math.floor((left + right) / 2);
         if (D[mid] < goal && goal < D[mid + 1] || D[mid] == goal) {
             return mid;
